+++ src/scss/index.scss
@@ -4,6 +4,7 @@
 
 @import 'parts/button';
 @import 'parts/code';
+@import 'parts/comment';
 @import 'parts/line';
 @import 'parts/loading';
 @import 'parts/header';
